+++ src/tasks/tslint/defaults.js
@@ -5,5 +5,6 @@ const path = require('path');
 module.exports = {
   configFile: path.join(__dirname, '../../../tslint.json'),
   extensions: ['.ts', 'tsx'],
-  files: require('../../utils/global-source-files')
+  files: require('../../utils/global-source-files'),
+  formatter: 'verbose'
 };