+++ src/tasks/webpack/helper-config.js
@@ -3,21 +3,11 @@
 const path = require('path');
 
 module.exports = {
-  getTSLoader: (tsConfig, babelConfig) => [
-    [
-      require.resolve('babel-loader'),
-      JSON.stringify(babelConfig)
-    ].join('?'),
-
-    // TODO: use query property of the loader
-    // awesome-typescript-loader does not support to specify babel package path yet
-    [
-      require.resolve('awesome-typescript-loader'),
-      JSON.stringify(Object.assign({
-        compiler: 'typescript',
-        tsconfig: path.join(__dirname, './tsconfig.json'),
-        resolveGlobs: false
-      }, tsConfig))
-    ].join('?')
-  ].join('!')
+  getTSLoader: (tsConfig, babelConfig) =>
+    Object.assign({
+      tsconfig: path.join(__dirname, './tsconfig.json'),
+      useBabel: true,
+      babelCore: require.resolve('babel-core'),
+      babelOptions: babelConfig
+    }, tsConfig)
 };